Split out Slawkenburgius’s Tale into two separate files
In the original print publication the first two pages of latin were laid out sentence-by-sentence on opposite pages, and this structure was retained for the Gutenberg publication. Given that most of our readers are going to be on single-page devices, I’ve reverted this to two separate chapters at the start of book four.